Added a features folder for use with behave testing
``behave'' is a beautiful Python module for writing acceptance tests
using what amounts to natural language (actually it's called Gherkin).

Please please please check out

  http://pythonhosted.org/behave/index.html

as we will be using is for the rest of the semester.

Once you have written the acceptance tests in Gherkin you link the
Gherkin statements up to actual code in the ``steps'' files, full of
@given, @when and @then decorators, associated Gherkin statements and
code.

The environment.py file is used to set up the environment for behave;
you can have it set up environments for each scenario, for the whole
``shooting match'' (as they are fond of saying in the behave
documentation for some reason), even for certain steps that have tags
you set.

As it stands, this commit represents only a framework; the steps
files(s) still needs the actual code inside the functions.  There also
remains some ambiguity about what exactly the requirements are (for
example, should users be able to edit their username?), but this is to
be expected.  This sprint consultation will be limited, but in the
next sprint hopefully such issues can be clarified.