Completed steps for first Scenario: View Current Details
The steps are all contained in features/steps/steps.py at this stage,
but as we use behave more there will likely have to be a better
organisational structure.

In my steps for this first Scenario I am assuming that there will be a
page called "/user/username" where users can edit their account
details.  The only thing I am expecting to definitely see on this page
at this stage is the string "Username: <username>", but as more
"details" are associated with a user there will be a growing list of
assertions required.