Cleaned up meterage.py
I have cleaned up some of the logic in meterage.py, and added some
docstrings:

https://www.python.org/dev/peps/pep-0257/

I would ask that everyone try to include docstrings, as it makes it
far clearer what the functions you write do for readers.

I have also added a couple of TODOs; in PyCharm, if your comment
starts with TODO then it gets picked up automatically if you go to
View > Tool Windows > TODO.  This is really useful for seeing the
minutiae of what needs to be done.  In this case, I said that some of
the funcitons rely on username being unique, and this is not
necessarily the case as far as I can tell.  Someone with a bit more
database experience needs to please remedy this.