added template_filter('newlines') to meterage.py
I don't know if it's just that nobody tried it, but if you actually
post to the site as it stood before this commit, you'll find that
newlines are not recognised and all the text simply gets displayed in
a one-line mess.

I have implemented a filer that goes right in the HTML using jinja2
(which comes packaged with Flask) to recognised the HTML linebreak
characters and send the text to a new line appropriately.