Separated config into files, including instance folder
For information on why I've done this, please see:

https://exploreflask.com/configuration.html

Usually an instance folder would not be committed to version control,
but for the purposes of this exercise I have done so.  I will use such
config files for, for example, defining the number of bcrypt rounds I
use in hashing the passwords.