Fixed all unit tests in meterage_tests.py
This was a fairly extensive overhaul.  Since users and passwords are
now stored in the database in the userPassword table, this meant that
in the set up for the tests we need to create the temporary database
and also initialise it with some users.  Just the admin and another
user are enough to see how the different accounts react under testing.

The tests were condensed down such that they automatically loop over
the database; this means that we don't need individual tests for
individual users, and indeed if you were to go into meterage_tests.py
and put more users into the database in the setup method you will find
that the tests will still all pass.

I also changed all instances of rv.data (i.e. accessing the data field
of the rv object) to rv.get_data().  The getter and setter methods are
preferred now in Flask, and the other method of simply accessing the
field is to be deprecated.  See:

A quote from the Flask API documentation
(http://flask.pocoo.org/docs/0.10/api/) says about 'data':

"
data
A descriptor that calls get_data() and set_data(). This should not be
used and will eventually get deprecated.
"

At the time of committing, all unit tests in meterage_tests.py are
passing.