Created User class with automatic hashing
*** At this stage I am still prototyping. ***

I have defined a User class with attributes 'username' and 'password'.
Upon initialising a User object the username is stored in a basic
fashion, but importantly the password is automatically hashed by the
flaskr_bcrypt extension.

I have set the number of Bcrypt rounds to twelve at this stage in
config.py, but this is open to future optimisation.

The requirements.txt file is there to allow you to automatically set
up a vitrual environment with the required Python packages.
flask_bcrypt is an extension from the approved extensions on the Flask
website, and ideally installation of the required packages using

  pip install -r requirements.txt

Will be nice and smooth.  We need to start thinking about how these
third party libraries are to be included in the repo; ideally using
this requirements.txt file will be sufficient.