Achieved functioning hashing and retrieval from the database
I am now able to hash the passwords and send the hashed password to the database.  This commit will require that you re-generate the database, easily achieved by simply deleting the existing database and running meterage.py as usual (I have included some code down the bottom to automatically generate the database if it doesn't find one where it expects to).

This branch is still a work in progress an requires much unit-testing and even probably a bit of a better implementation.  It is quite clunky at the moment due to the lack of an ORM, but by next sprint we will ideally be using SQLAlchemy to smooth this over.