Incorporated new test, changed database	initialisation
Gravatar tests have been condensed into one test looping over a list
of emails addresses.  The email addresses are being posted to the
/login page, but at this stage meterage.py does not include anything
that handles any email address posted to it.

I removed the unused method definition at the bottom of the page as it
is a replica of another function present in meterage.py

The test is not a test of Gravatar functionality, only that the email
address posted is returned in the results of the post.