behave code written
I have writte behave code for the main Scenarios:

1.  Making a post
2.  Managing my details
3.  Removing my Gravatar assignment.

Scenario 1 is completely passing, so I consider this grounds for a
pull request.

Scenario 2 and 3 are not passing, but this is because this is to do
with "managing details".  In fact I have not even written one of the
steps for "removing my Gravatar assignment", as it is too contingent
on how the person that implements the managing details web interface
chooses to do it.  I suggested a Boolean column in the userPassword
table that sets whether or not the Gravatar is shown.

I was forced to update some of the code in the old steps functions as
they are rapidly going out of date with the changing program.  For
example, the steps were previously drawing from the static dictionary
that had usernames and passwords in it from the old flaskr.py.  I have
tweaked the methods to work again, but any further tweaking would be
pointless until people actually implement their user stories and we
get things merged properly into the mastr branch.