Began to implement user web interface
By clicking on your username towards the top of layout.html you are
taken to a page to manage your details.  At this point, this page is
simply displaying username and Gravatar email address.

I was the user to be able to change their username, and also to be
able to change the Gravatar email address associated with their
account.

Users should also be able to change their password.