Login, main page working except for Gravatars
I need to fix up the definition of the Entry class such that when I
obtain the user_id, I can use this to obtain the gravataremail
associated with this user_id in the users table and therefore pass
this to the function to generate the Gravatar itself.

At this stage, my lack of experience withdatabases is letting me
down.  I'm fairly sure I want to perform a join on the used_id between
entries and users tables, then somehow extract the Gravatar.  I can do
this easily with a query string, but I want this to be automatic.
When I ask for all the entries, the gravatar should come as part of
the request automatically.