Near full functionality
The code needs a refactor and some explanatory commenting, but save
for this it is almost fully functional.

The only place that things are not fully functional is in the Gravatar
department.  I can easily get it working, but the implementation would
be somewhat of a hack.  I would prefer the Gravatart email to be a a
column in the Entry class that matches up with the User class via the
user id of the entry poster.