Wrote semi-functional setup and teardown methods
The set-up method sets up some config parameters in the
meterage.app.config dictionary and then reloads the meterage.models
script, which contains all of the models used in the database.  We
need to perform this reload as all of the models need to inherit the
db.Model object that we have just set up in our set up method.

At the moment we are able to pass the very first test, but there is an
issue with the next setup after the teardown.  This needs to be
resolved, but this commit represents a decent step forward; worthy of
a commit.