Unit tests passing
All of the unit tests that were previously passing (i.e. not those
that raise a NotImplementedError) are passing now.  I am afraid that
in order ot achieve this I was forced to use somewhat of a hack; I
disabled DEBUG mode.  I think that Flask does not like SQLAlchemy
during testing, though I am not sure.

I still need to clean up the code considerably and annotate it, but at
the moment the unit tests are passing.