Fixed managing details page
The page had some issues with a Javascript method I was using to
toggle buttons being displayed.  I have used a Bootstrap modal box to
replace the existing functionality.  Now a user can click the 'edit
details' button and is presented with a modal pop-up in which to edit
their details.  Should they click save, the details will be changed.