All ORM unit tests passing
For some reason, at this point the ORM tests refuse to all pass when
running as part of the test suite.  This has something to do with
SQLAlchemy.  Perhaps something to do with the Session object, or
threads?

In any case, the tests *are* passing and this means that we are able
to proceed.