All unit tests passing
Note that some of the unit tests do not work when the whole suite is
run, they only work individually.  I am not sure why this is at this
stage, and it certainly needs to be understood and ironed out in the
future, but for now the tests are passing and we can move on.

The issue is something to do with the way SQLAlchemy works.  Even with
Waiho helping we could not figure it out.

It may also be linked to the fact that we have had to make DEBUG False
in our config to get the tests to run.